making the panel a separate component that is nested within the Canvas component
