+++ src/components/CSS/InputPanel.css
@@ -1,0 +1,33 @@
+
+#inputPanel {
+    position: fixed;
+    bottom: 0;
+    left: 0;
+    margin-left: 1vw;
+    margin-bottom: 1vw;
+    background-color: white;
+}
+
+#inputPanel > div {
+    padding: 15px 10px;
+    display: block;
+}
+
+#inputPanel > div > input {
+    width: 50px;
+    padding: 10px;
+    margin-left: 6px;
+    margin-right: 3px;
+
+    outline-style: none;
+    box-shadow: none;
+    border-color: transparent;
+
+    border: black 2px solid;
+}
+
+#inputPanel > div > input:focus {
+
+    background-color: rgb(238, 251, 255);
+    border: blue 2px solid;
+}
