+++ src/components/Canvas.js
@@ -24,6 +24,7 @@ import { findPivotGlobalCoords } from "../processingFunctions/findPivotGlobalCoo
 import { findCannonTopLeftGlobalCoords, topLeftConerVelocityBar } from "../processingFunctions/topLeftCorners";
 import { calclateGrowthFactorVelocity } from "../processingFunctions/calculateGrowthFactor";
 import FireButton from "./FireButton";
+import InputPanel from "./InputPanel";
 
 export default function Canvas() {
 
@@ -51,7 +52,7 @@ export default function Canvas() {
   const velocitySliderInfo = getHolsterInfo("velocity_slider");
   
   const [elevationAngle, setElevationAngle] = useState(0);
-  const [launchVelocity, setLaunchVelocity] = useState(0)
+  const [launchVelocity, setLaunchVelocity] = useState(10)
 
   // User state variables
   const cannonClick = useRef(false);
@@ -82,7 +83,15 @@ export default function Canvas() {
   useEffect(() => {
     ctxRef.current = canvasRef.current.getContext('2d');
     drawDefaultCannon(ctxRef.current, canvasRef.current, cannonRef.current, holsterRef.current, cannonInfo, holsterInfo, USER_ANCHOR_POINT.current)
-    drawDefaultVelocitySlider(ctxRef.current, canvasRef.current, velocityBarRef.current, velocitySliderRef.current, findCannonTopLeftGlobalCoords(canvasRef.current, USER_ANCHOR_POINT.current, cannonInfo))
+    drawDefaultVelocitySlider(
+      ctxRef.current, 
+      canvasRef.current, 
+      velocityBarRef.current, 
+      velocitySliderRef.current, 
+      findCannonTopLeftGlobalCoords(canvasRef.current, USER_ANCHOR_POINT.current, cannonInfo), 
+      MAX_SPEED, 
+      launchVelocity
+    )
   })
 
   useEffect(() => {
@@ -123,53 +132,6 @@ export default function Canvas() {
     ctxRef.current.stroke();
     ctxRef.current.fill();
   }, [cannonInfo, elevationAngle, holsterInfo, width, launchVelocity])
-
-  //////////////////////////////// Textbox Input ////////////////////////////////////////////////////
-
-  function changeAngleWithTextBox(e) {
-    const val = e.target.value;
-    // requires some defensive programming
-    try {
-      if (val === "") {
-        setElevationAngle(0);
-      }
-      // some defensive programming
-      if (isNaN(parseFloat(val))) {
-        return;
-      } else if (parseFloat(val) < 0) {
-        setElevationAngle(0)
-      } else if (parseFloat(val) > 90) {
-        setElevationAngle(90);
-      } else {
-        setElevationAngle(parseFloat(val))
-      }
-    } catch (error) {
-      console.error("In Canvas.js | function changeAngleWithTextBox")
-      console.error(error.message)
-      return;
-    }
-  }
-
-  function changeVelocityWithTextBox(e) {
-    const val = e.target.value;
-    try {
-      if (val === "") {
-        setLaunchVelocity(0);
-      }
-      if (isNaN(parseFloat(val))) {
-        return;
-      } else if (parseFloat(val) < 0) {
-        setLaunchVelocity(0);
-      } else if (parseFloat(val) > MAX_SPEED) {
-        setLaunchVelocity(MAX_SPEED);
-      } else {
-        setLaunchVelocity(parseFloat(val));
-      }
-    } catch (error) {
-      console.error("In Canvas.js | function changeVelocityWithTextBox")
-      console.error(error.message)
-    }
-  }
 
   //////////////////////// Changing Angles Mouse Events ////////////////////////
 
@@ -225,8 +187,6 @@ export default function Canvas() {
       const mouse_y = e.pageY;
 
       const xDisplacement = (mouse_x  - click_x.current) * window.devicePixelRatio;
-      // TODO: 817 is a magic number that needs to be better handled throughout the ENTIRE code base
-      // 817 refers to the width of the velocity bar in pixels
       const velocityPerPixel = MAX_SPEED / (velocitySliderInfo.pixel_width * calclateGrowthFactorVelocity(canvasRef.current));
       
       click_x.current = mouse_x;
@@ -331,29 +291,14 @@ export default function Canvas() {
         />
       </canvas>
 
-      <div id="inputPanel">
-        <div id="velocityInput">
-          Velocity:
-          <input 
-            type="text"
-            ref={velocityInputRef}
-            onChange={(e) => changeVelocityWithTextBox(e)}
-            maxLength={8}
-          />
-          m/s
-        </div>
-        <div id="angleInput">
-          Angle: 
-          <input 
-            type="text" 
-            ref={angleInputRef}
-            onChange={(e) => {changeAngleWithTextBox(e)}} 
-            style={{bottom: "95px"}}
-            maxLength={6}
-          />
-          degrees
-        </div>
-      </div>
+      <InputPanel 
+        setElevationAngle={setElevationAngle} 
+        setLaunchVelocity={setLaunchVelocity} 
+        MAX_SPEED={MAX_SPEED} 
+        angleInputRef={angleInputRef} 
+        velocityInputRef={velocityInputRef}
+      />
+
       <FireButton fireCannon={fireCannon} />
     </>
